fix: use chain service
