+++ src/renderer/features/multisig-wallet-create/multisig-wallet/model/chain-model.ts
@@ -3,10 +3,10 @@ import { BN } from '@polkadot/util';
 import { combine, createEffect, createStore, sample } from 'effector';
 import { spread } from 'patronum';
 
+import { chainsService } from '@/shared/api/network';
 import { type Asset, type Chain, type Transaction } from '@/shared/core';
 import { series } from '@/shared/effector/series';
 import { TEST_ACCOUNTS, getNativeAsset, merge, nonNullable, nullable, withdrawableAmountBN } from '@/shared/lib/utils';
-import { sortChains } from '@/shared/lib/utils/chains';
 import { accountService } from '@/domains/network';
 import { balanceModel, balanceUtils } from '@/entities/balance';
 import { networkModel, networkUtils } from '@/entities/network';
@@ -28,7 +28,7 @@ type ChainsWithFee = {
 
 const $multisigChains = combine(networkModel.$chains, chains => {
   const filteredChains = Object.values(chains).filter(chain => networkUtils.isMultisigSupported(chain.options));
-  return sortChains(filteredChains);
+  return chainsService.sortChains(filteredChains);
 });
 
 const calculateFeeForChainFx = createEffect(
