+++ src/renderer/features/proxied-add-pure/model/form-model.ts
@@ -1,6 +1,7 @@
 import { BN } from '@polkadot/util';
 import { combine, createEvent, createStore, restore, sample } from 'effector';
 
+import { chainsService } from '@/shared/api/network';
 import { type Chain, type ProxiedAccount, type Transaction, TransactionType, type Wallet } from '@/shared/core';
 import { type Form, createForm } from '@/shared/forms';
 import {
@@ -10,7 +11,6 @@ import {
   transferableAmountBN,
   withdrawableAmountBN,
 } from '@/shared/lib/utils';
-import { sortChains } from '@/shared/lib/utils/chains';
 import {
   createComplexTxStore,
   createInitiatorsStore,
@@ -171,7 +171,7 @@ const $availableChains = combine(
     const filteredChains = proxyChains.filter(chain => {
       return walletAccounts.some(account => accountService.isAccountAvailableOnChain(account, chain));
     });
-    return sortChains(filteredChains);
+    return chainsService.sortChains(filteredChains);
   },
 );
